Aggiunte diverse funzionalità

+++ src/main/java/it/unibo/application/dto/Misurazione.java
@@ -1,9 +1,10 @@
 package it.unibo.application.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record Misurazione(
     String username,
     LocalDate data,
-    int peso
+    BigDecimal peso
 ) {}
